% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_progress_comment.R
\name{add_progress_comment}
\alias{add_progress_comment}
\title{add progress comment}
\usage{
add_progress_comment(text)
}
\arguments{
\item{text}{comment to add to progress file}
}
\description{
adds a comment to the progress file of standard format
}
