% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_progress_file.R
\name{create_progress_file}
\alias{create_progress_file}
\title{Progress file creator}
\usage{
create_progress_file(path, name = NULL)
}
\arguments{
\item{path}{path within which the file is created}

\item{name}{name of file. Defaults to date_time_progress.txt}
}
\description{
Creates a progress file in a desired location
}
\details{
Creates a progress file in a desired location and loads the name of the file into the global environment. Allows for the progress of long scripts to be tracked by adding updates into the file.
}
\author{
Daniel Padfield
}
