% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eilers_Inhib.R
\name{Eilers_Inhib}
\alias{Eilers_Inhib}
\title{Eilers PI curve inhibtion ratio}
\usage{
Eilers_Inhib(Pmax, Iopt, a)
}
\arguments{
\item{Pmax}{The photosynthetic maxima}

\item{Iopt}{The optimal light intensity}

\item{a}{The initial slope of the curve}
}
\description{
Gives the ratio of actual Pmax and the Pmax that would occur if no inhibition occurred.
}
\references{
Eilers, P.H.. & Peeters, J.C.. (1988) A model for the
relationship between light intensity and the rate of photosynthesis in
phytoplankton. Ecological Modelling, 42, 199–215.
}
\seealso{
\code{\link{Eilers_PI}} and \code{\link{Eilers_mod_PI}} for fitting the Eilers photosynthesis irradiance curve
}
\author{
Daniel Padfield
}
