% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCR_mix.R
\name{PCR_mix}
\alias{PCR_mix}
\title{PCR mix template}
\usage{
PCR_mix(reps, extra = 2)
}
\arguments{
\item{reps}{the number of PCRs being run for a single primer set}

\item{extra}{the number of extras you want to calculate to account for pipetting error - defaults to 2}
}
\description{
Produces a table of the amounts needed for a standard 10µl PCR mix with 1µl of DNA template.
}
\examples{
PCR_mix(14, 1)
PCR_mix(14,2)
PCR_mix(14)

}
\author{
Daniel Padfield
}
