% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Platt_PI.R
\name{Platt_PI}
\alias{Platt_PI}
\title{Fitting a photosynthesis-irradiance curve}
\usage{
Platt_PI(I, ps, alpha, beta)
}
\arguments{
\item{I}{The value of light, either a vector or a single, numeric value.}

\item{ps}{Scaling coefficient for the curve}

\item{alpha}{The steepness of the initial gradient of the curve - is a
measure of photosynthetic efficiency}

\item{beta}{The steepness and extent of photoinhibition as light levels
increase past the optimum}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Uses the Platt PI model with photoinhibition (Platt 1990) to fit a model to
photosynthesis data across many light levels typically taken using the
Oxylab system
}
\details{
When using nlsLoop, set the parameters lower limits to 0 as the equation has
two solutions and with equivalent AIC scores so could return some unwanted
negative values for some parameters.
}
\note{
%% ~~further notes~~
}
\examples{



}
\author{
Daniel Padfield
}
\references{
Platt, T., Sathyendranath, S. & Ravindran, P. (1990) Primary
production by phytoplankton : analytic solutions for daily rates per unit
area of water surface. Proceedings: Biological Sciences, 241, 101–111.
}
\seealso{
\code{\link{Eilers_PI}} for fitting an alternative photosynthesis
irradiance curve. \code{\link{Pmax}} for getting the maximal photosynthetic
rate of the Platt fitted curve.
}
\keyword{~kwd1}
\keyword{~kwd2}

