% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schoolfield_Topt.R
\name{Schoolfield_Topt}
\alias{Schoolfield_Topt}
\title{Calculates the optimum temperature}
\usage{
Schoolfield_Topt(Eh, Th, Ea, K = c("Y", "N"))
}
\arguments{
\item{Eh}{The value of the deactivation energy}

\item{Th}{The value of Th being the temperature at which the substrate is
half high temperature suppressed (in K)}

\item{Ea}{The activation energy}

\item{K}{Whether or not you want the results in K or in degrees centigrade.
Defaults to K. 'Y' gives K and 'N' gives degrees celsius}
}
\value{
value of optimum temperature
}
\description{
Calculates the optimum temperature from the parameters of the modified
Sharpe-Schoolfield equation. Used in tandem with schoolfield.high
}
\examples{
Schoolfield_Topt(4, 315, 0.6)
}
\author{
Daniel Padfield
}
