\name{TeamPhytoplankton}
\alias{TeamPhytoplankton}
\alias{TeamPhytoplankton}
\docType{package}
\title{
TEAM PHYTOPLANKTON FUNCTIONS
}
\description{
A place for all Team Phytoplankton functions to make our lives a little easier
}
\details{
Just thought I would put all the functions Team Phytoplankton use regularly into a fancy little package. Is a work in progress but currently the functions included are:

Boltzmann()

schoolfield.high()

std.error()

read_data() - for O2 view files

Pmax()

Platt_PI()

Topt()
}
\author{
Daniel Padfield

Maintainer: Daniel Padfield
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
