% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamPhytoplankton-package.R
\docType{package}
\name{TeamPhytoplankton}
\alias{TeamPhytoplankton}
\alias{TeamPhytoplankton-package}
\title{Team Phytoplankton Functions}
\description{
A place for all Team Phytoplankton functions to make our lives a little
easier
}
\details{
Just thought I would put all the functions Team Phytoplankton use regularly
into a fancy little package. Is a work in progress but currently the
functions included are:

boltzmann()

schoolfield.high()

std.error()

read_data() - for O2 view files

Pmax()

Platt_PI()

Topt()

FSCtoMicron()

inv.temp()

quasi.rsq.nls()

Eilers_PI()

nlsLoop()
}
\examples{

~~ simple examples of the most important functions ~~

}
\seealso{
GitHub link - https://github.com/padpadpadpad/TeamPhytoplankton
}
\author{
Daniel Padfield

Maintainer: Daniel Padfield - d.padfieldscfc@gmail.com
}
\keyword{EVOLUTION,}
\keyword{MODELLING}
\keyword{NONLINEAR}
\keyword{PLANKTON,}
