% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tmax.R
\name{Thomas.Tmax}
\alias{Thomas.Tmax}
\title{Calculates Tmax for the Thomas TRC}
\usage{
Thomas.Tmax(Topt, w)
}
\arguments{
\item{Topt}{The parameter Tmax}

\item{w}{The parameter w}
}
\value{
returns the Tmax for the Thomas TRC. The highest temperature at
which the rate is no longer positive.
}
\description{
Calculates the Tmax for the thermal response curve by adding w/2 onto Topt.
}
\note{
%% ~~further notes~~
}
\examples{



}
\references{
Thomas, M.K., Kremer, C.T., Klausmeier, C. a & Litchman, E.
(2012) A global pattern of thermal adaptation in marine phytoplankton.
Science, 338, 1085–8.

Thomas, M.K., Kremer, C.T. & Litchman, E. (2015) Environment and
evolutionary history determine the global biogeography of phytoplankton
temperature traits. Global Ecology and Biogeography, 25, 75-86.
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Daniel Padfield
}
\keyword{~kwd1}
\keyword{~kwd2}
