\name{Thomas.Tmax}
\alias{Thomas.Tmax}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculates Tmax for the Thomas TRC
}
\description{
Calculates the Tmax for the thermal response curve by adding w/2 onto Topt.
}
\usage{
Tmax(Topt, w)
}
\arguments{
  \item{Topt}{
The parameter Tmax
}
  \item{w}{
The parameter w
}
}
\details{

}
\value{
returns the Tmax for the Thomas TRC. The highest temperature at which the rate is no longer positive.
}
\references{
Thomas, M.K., Kremer, C.T., Klausmeier, C. a & Litchman, E. (2012) A global pattern of thermal adaptation in marine phytoplankton. Science, 338, 1085–8.

Thomas, M.K., Kremer, C.T. & Litchman, E. (2015) Environment and evolutionary history determine the global biogeography of phytoplankton temperature traits. Global Ecology and Biogeography, 25, 75-86.
}
\author{
Daniel Padfield
}
\note{
%%  ~~further notes~~
}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
