% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eilers_PI.R
\name{Eilers_PI}
\alias{Eilers_PI}
\title{Fitting a photosynthesis-irradiance curve}
\usage{
Eilers_PI(Pmax, Iopt, a, I)
}
\arguments{
\item{Pmax}{The photosynthetic maxima}

\item{Iopt}{The optimal light intensity}

\item{a}{The initial slope of the curve}

\item{I}{Light}
}
\description{
Uses the Eilers & Peeters (1988) PI model to fit a model to photosynthesis
data across many light levels typically taken using the Oxylab system.
Benefits from modelling Pmax without any other derivatives so can test the
probability of the parameter.
}
\details{
When using nlsLoop, set the parameters lower limits to 0 as the equation has
two solutions and with equivalent AIC scores so could return some unwanted
negative values for some parameters.
}
\examples{



}
\references{
Eilers, P.H.. & Peeters, J.C.. (1988) A model for the
relationship between light intensity and the rate of photosynthesis in
phytoplankton. Ecological Modelling, 42, 199–215.
}
\seealso{
\code{\link{Platt_PI}} for fitting an alternative photosynthesis
irradiance curve
}
\author{
Daniel Padfield
}
