% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSCtoMicron.R
\name{FSCtoMicron}
\alias{FSCtoMicron}
\title{Converts FSC.H to micrometres}
\usage{
FSCtoMicron(x)
}
\arguments{
\item{x}{vector of values of FSC.H to convert}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Uses the results of a linear regression to convert values of FSC.H on the
ESI BD Accuri C6 to values of micrometres
}
\details{
Uses an equation of µm = 1.036827e-05*FSC.H + 1.041811

Uses raw values of FSC.H before it has been log transformed.

The graph of this regression is in the shared Dropbox and has an r squared
value of 0.78. So it isn't perfect but a very good approximate.
}
\examples{

data <- data.frame(length = c(1,  2,  4, 6, 10, 15), FSC.H = c(58127.95,   72891.38,  339312.54,  619733.87, 1096084.37, 1767728.37))

FSCtoMicron(data$FSC.H)


}
\author{
Daniel Padfield
}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}

