% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard error.R
\name{std.error}
\alias{std.error}
\title{Calculates the standard error}
\usage{
std.error(x, na.action = c(FALSE, TRUE))
}
\arguments{
\item{x}{vector of values to calculate standard error for (must be numeric)}

\item{na.action}{whether there are NAs in the data or not. Defaults to
FALSE. TRUE means that both the row including NA is excluded so both the
number of samples and the mean is affected}
}
\value{
%% ~Describe the value returned %% If it is a LIST, use %%
}
\description{
Calculates the standard error but takes into account whether NAs are present
or not
}
\note{
%% ~~further notes~~
}
\examples{

y <- c(1,4,6,7,91,3,45,6,7,8,345,232,43423,5436, NA)

std.error(y)
NA
std.error(y, TRUE)
3091.521


}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Daniel Padfield
}
\keyword{~kwd1}
\keyword{~kwd2}
