% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockSolVol.R
\name{StockSolVol}
\alias{StockSolVol}
\title{Calculates the volume of stock solution needed when making up working
solutions.}
\usage{
StockSolVol(StockSolConc, NewSolConc, NewSolVol)
}
\arguments{
\item{StockSolConc}{The concentration of your stock solution}

\item{NewSolConc}{The desired concentration of your new solution}

\item{NewSolVol}{The total volume of your new solution}
}
\description{
Uses the equation:
}
\details{
StockSolVol = (NewSolConc * NewSolVol) / StockSolConc

To work out the volume of your original solution (stock solution) you need
to make up the desired concentration of your new solution (eg. for a working
solution)
}
\examples{

StockConc <- 100
DesiredConc <- 10
DesiredVol <- 50

StockSolVol(StockConc, DesiredConc, DesiredVol)


}
\author{
Daniel Padfield
}
\keyword{~kwd1}
\keyword{~kwd2}

